Issue #33: moved get_field_widths, display_header, and display_status codes to wherepy/io/__init__.py
